fix: address circular deps
