+++ packages/ibm-products/src/components/Datagrid/Datagrid/addons/Filtering/hooks/useFilters.js
@@ -32,35 +32,7 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 import OverflowCheckboxes from '../OverflowCheckboxes';
 import { getInitialStateFromFilters } from '../utils';
 import { usePreviousValue } from '../../../../../../global/js/hooks';
-
-export const handleCheckboxChange = ({
-  checked,
-  filtersState,
-  column,
-  option,
-  setFiltersState,
-  applyFilters,
-  type,
-}) => {
-  const checkboxCopy = filtersState[column].value;
-  const foundCheckbox = checkboxCopy.find(
-    (checkbox) => checkbox.value === option.value
-  );
-  foundCheckbox.selected = checked;
-  setFiltersState({
-    ...filtersState,
-    [column]: {
-      value: checkboxCopy,
-      type,
-    },
-  });
-  applyFilters({
-    column,
-    value: [...filtersState[column].value],
-    type,
-  });
-  option.onChange?.(checked);
-};
+import { handleCheckboxChange } from '../handleCheckboxChange';
 
 const useFilters = ({
   updateMethod,
