+++ packages/ibm-products/src/components/PageHeader/PageHeaderTitle.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { SkeletonText } from '@carbon/react';
-import { EditInPlace } from '../';
+import { EditInPlace } from '../EditInPlace';
 
 /**
  *
