+++ packages/ibm-products/src/components/Saving/Saving.test.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2021, 2021
+ * Copyright IBM Corp. 2021, 2024
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -9,8 +9,7 @@ import { render, screen, act } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
-import { Saving } from '.';
-console.log(Saving);
+import { Saving } from './Saving';
 
 const componentName = Saving.displayName;
 const defaultProps = {
